+++ src/function/relational/compareNatural.js
@@ -76,6 +76,7 @@ export const createCompareNatural = /* #__PURE__ */ factory(name, dependencies, 
    *
    * History:
    *
+   *     v15.2   Handle bigints
    *     v3.14   Created
    *
    *
@@ -92,8 +93,8 @@ export const createCompareNatural = /* #__PURE__ */ factory(name, dependencies, 
     let c
 
     // numeric types
-    if ((typeX === 'number' || typeX === 'BigNumber' || typeX === 'Fraction') &&
-        (typeY === 'number' || typeY === 'BigNumber' || typeY === 'Fraction')) {
+    if ((typeX === 'number' || typeX === 'BigNumber' || typeX === 'Fraction' || typeX === 'bigint') &&
+        (typeY === 'number' || typeY === 'BigNumber' || typeY === 'Fraction' || typeY === 'bigint')) {
       c = compare(x, y)
       if (c.toString() !== '0') {
         // c can be number, BigNumber, or Fraction