+++ HISTORY.md
@@ -2,7 +2,8 @@
 
 # unreleased changes since 15.1.0
 
-- Doc: Correct several arithmetic and relational documentation examples
+- Fix: #3631 Handle bigints in `compareNatural` (#3632). Thanks @Dheemanth07.
+- Docs: Correct several arithmetic and relational documentation examples
   and add History (#3630). Thanks @Anadian.
 - Fix: #3578 interpret empty true-expr of conditional as error (#3581).
   Thanks @gwhitney.